moved CHECK_PARAMS_SYNTAX_REGEX
