+++ wiggle_query_language/clauses/parsing_helpers/parse_statement_checks.py
@@ -11,10 +11,13 @@ from exceptions.wql.parsing import (
 from wiggle_query_language.clauses.regexes.make.make_patterns import (
     ILLEGAL_CHARS_REGEX,
     MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX,
-    MAKE_STATEMENT_CHECK_PARAMS_SYNTAX_REGEX,
+    # CHECK_PARAMS_SYNTAX_REGEX,
     PARAM_LIST_VALUE_REGEX,
 )
 
+from wiggle_query_language.clauses.regexes.patterns.properties import (
+    CHECK_PARAMS_SYNTAX_REGEX,
+)
 from wiggle_query_language.clauses.regexes.patterns.relationships import (
     RELATIONSHIP_DIR_CHECK_REGEX,
 )
@@ -43,7 +46,7 @@ def check_node_rel_properties(stmt_matches: list[str]) -> True:
     """
 
     for stmt in stmt_matches:
-        if not (param_string := MAKE_STATEMENT_CHECK_PARAMS_SYNTAX_REGEX.findall(stmt)):
+        if not (param_string := CHECK_PARAMS_SYNTAX_REGEX.findall(stmt)):
             continue
 
         for param_match_in in param_string:
