+++ wiggle_query_language/clauses/regexes/__init__.py
@@ -1,0 +1,2 @@
+EXTRA_ALLOWED_CHARS = "@"
+ILLEGAL_CHARS = "#%&*$"
