+++ wiggle_query_language/clauses/regexes/make/tests/test_make_re.py
@@ -7,7 +7,6 @@ from wiggle_query_language.clauses.regexes.make.make_patterns import (
     ILLEGAL_CHARS_REGEX,
     MAKE_STATEMENT_ALL_REGEX,
     MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX,
-    MAKE_STATEMENT_CHECK_PARAMS_SYNTAX_REGEX,
     NODES_RELS_PATTERN_REGEX,
     PARAM_LIST_VALUE_REGEX,
 )
@@ -91,32 +90,6 @@ class TestMakeRePatterns:
     ) -> None:
         with exception:
             test = MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX.findall(test_pattern)
-            assert test == expected_result
-
-    @pytest.mark.parametrize(
-        "test_pattern, expected_result, exception",
-        [
-            pytest.param(
-                """MAKE (:NodeLabel{int: 1, str: '2', str2:"2_4", float: 3.14, bool: true, list: [1, '2', "2_4", "3 4", 3.14]});""",
-                [
-                    """{int: 1, str: '2', str2:"2_4", float: 3.14, bool: true, list: [1, '2', "2_4", "3 4", 3.14]}"""
-                ],
-                does_not_raise(),
-                id="EXP PASS: 1 Match",
-            ),
-            pytest.param(
-                "REPORT (:NodeLabel);",
-                [],
-                does_not_raise(),
-                id="EXP PASS: No Match",
-            ),
-        ],
-    )
-    def test_make_statement_check_params_syntax(
-        self, test_pattern: str, expected_result: Optional[list[str]], exception
-    ) -> None:
-        with exception:
-            test = MAKE_STATEMENT_CHECK_PARAMS_SYNTAX_REGEX.findall(test_pattern)
             assert test == expected_result
 
     @pytest.mark.parametrize(
